Show default help message for "How can Amazon Q help me?"
