+++ server/aws-lsp-codewhisperer/src/language-server/constants.ts
@@ -1,0 +1,3 @@
+export const MISSING_BEARER_TOKEN_ERROR = 'credentialsProvider does not have bearer token credentials'
+export const INVALID_TOKEN = 'The bearer token included in the request is invalid.'
+export const GENERIC_UNAUTHORIZED_ERROR = 'User is not authorized to make this call'
